Add aria-sort attribute to the appropriate th
Add the aria-sort attribute to the th of the column by which the data is
sorted. This is conceivably useful to let screen readers know how the
table is sorted, but it also allows CSS styles to be applied, for
example to highlight the sorted column header or to change an icon in
the column header to indicate the sort direction.

This required breaking the start of the table into separate variables so
that the column headers can be addressed individually. This is a step
toward being able to reorder (#66) or remove (#133) columns.