Add sort criterion and order classes to table tag
Add a class "sort-col-1", "sort-col-2", or "sort-col-3" to the table tag
to indicate by which column the data is sorted, and a class "asc" or
"desc" to indicate whether it is sorted in ascending or descending
order. This is intended to be useful for applying a style to the sorted
column or its header.

Modify the default template to slightly darken the sorted column.